[SOIN] Extrais un composant pour la description longue d’une mesure
...qui peut être repliée
