+++ svelte/lib/mesure/contenus/ContenuOngletMesure.svelte
@@ -4,6 +4,7 @@
   import { validationChamp } from '../../directives/validationChamp';
   import { rechercheTextuelle } from '../../tableauDesMesures/stores/rechercheTextuelle.store';
   import type { ReferentielStatut } from '../../ui/types';
+  import DescriptionLongueMesure from '../../ui/DescriptionLongueMesure.svelte';
 
   export let estLectureSeule: boolean;
   export let categories: Record<string, string>;
@@ -35,20 +36,10 @@
     </label>
   {/if}
   {#if $configurationAffichage.doitAfficherDescriptionLongue}
-    <details open={true}>
-      <summary />
-      <p>
-        {@html texteSurligne}
-      </p>
-      {#if $store.mesureEditee.mesure.lienBlog}
-        <a
-          class="lien-blog"
-          href={$store.mesureEditee.mesure.lienBlog}
-          target="_blank"
-          rel="noopener">Comment mettre en œuvre cette mesure ?</a
-        >
-      {/if}
-    </details>
+    <DescriptionLongueMesure
+      description={texteSurligne}
+      lienBlog={$store.mesureEditee.mesure.lienBlog}
+    />
   {/if}
 
   <div class="conteneur-statut">
@@ -155,7 +146,6 @@
     margin-right: 4px;
   }
 
-  details,
   .conteneur-statut,
   label {
     margin-bottom: 30px;
@@ -183,52 +173,5 @@
 
   .mention-requis .asterisque {
     color: var(--rose-anssi);
-  }
-
-  summary {
-    font-weight: bold;
-    cursor: pointer;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-  }
-
-  summary:before {
-    content: 'Description de la mesure';
-  }
-
-  summary:after {
-    content: '';
-    width: 24px;
-    height: 24px;
-    background: url('/statique/assets/images/chevron_noir.svg');
-    transform: translateY(1px) rotate(90deg);
-    transition: transform 0.1s ease-in-out;
-  }
-
-  details[open] > summary:after {
-    transform: translateY(1px) rotate(270deg);
-  }
-
-  details p {
-    margin: 8px 0 0;
-    font-weight: 400;
-  }
-
-  a.lien-blog {
-    display: block;
-  }
-
-  a.lien-blog::after {
-    display: inline-block;
-    content: '';
-    width: 16px;
-    height: 16px;
-    background: url('/statique/assets/images/home/icone_lien_externe.svg')
-      no-repeat center;
-    filter: invert(29%) sepia(71%) saturate(1558%) hue-rotate(181deg)
-      brightness(70%) contrast(132%);
-    margin: auto 5px;
-    background-position-y: 2px;
   }
 </style>
